+++ Snowflake.Data.Tests/UnitTests/ChunkDeserializerTest.cs
@@ -2,7 +2,7 @@
  * Copyright (c) 2021 Snowflake Computing Inc. All rights reserved.
  */
 
-namespace Snowflake.Data.Tests
+namespace Snowflake.Data.Tests.UnitTests
 {
     using Newtonsoft.Json;
     using NUnit.Framework;
